Filenames with white spaces now supported
Currently it is not possible to call the list method and then
pipe the entries of a RAR file if the RAR file has white spaces in
its file name (or path). Appending and prepending the `"`
character when calling the UnRar constructor does not help either
since it only works for the list method but fails when calling
pipe.